Address most of PJB's feedback

+++ Content.Shared/Atmos/Atmospherics.cs
@@ -173,7 +173,7 @@ namespace Content.Shared.Atmos
         /// <summary>
         ///     Total number of gases. Increase this if you want to add more!
         /// </summary>
-        public const int TotalNumberOfGases = 6;
+        public const byte TotalNumberOfGases = 6;
 
         /// <summary>
         ///     Amount of heat released per mole of burnt hydrogen or tritium (hydrogen isotope)
